Python3: split up property bindings
